package refactor from mil.nga.giat.wkb to mil.nga.wkb
