+++ src/main/java/mil/nga/wkb/geom/GeometryEnvelope.java
@@ -1,4 +1,4 @@
-package mil.nga.giat.wkb.geom;
+package mil.nga.wkb.geom;
 
 /**
  * Geometry envelope
