+++ src/main/java/mil/nga/wkb/geom/LineString.java
@@ -1,4 +1,4 @@
-package mil.nga.giat.wkb.geom;
+package mil.nga.wkb.geom;
 
 import java.util.ArrayList;
 import java.util.List;
