+++ src/main/java/mil/nga/wkb/geom/MultiPoint.java
@@ -1,4 +1,4 @@
-package mil.nga.giat.wkb.geom;
+package mil.nga.wkb.geom;
 
 import java.util.List;
