+++ src/main/java/mil/nga/wkb/geom/MultiSurface.java
@@ -1,12 +1,13 @@
-package mil.nga.giat.wkb.geom;
+package mil.nga.wkb.geom;
 
 /**
  * A restricted form of GeometryCollection where each Geometry in the collection
- * must be of type Curve.
+ * must be of type Surface.
  * 
  * @author osbornb
  */
-public abstract class MultiCurve<T extends Curve> extends GeometryCollection<T> {
+public abstract class MultiSurface<T extends Surface> extends
+		GeometryCollection<T> {
 
 	/**
 	 * Constructor
@@ -15,7 +16,7 @@ public abstract class MultiCurve<T extends Curve> extends GeometryCollection<T> 
 	 * @param hasZ
 	 * @param hasM
 	 */
-	protected MultiCurve(GeometryType type, boolean hasZ, boolean hasM) {
+	protected MultiSurface(GeometryType type, boolean hasZ, boolean hasM) {
 		super(type, hasZ, hasM);
 	}
