+++ src/main/java/mil/nga/wkb/io/WkbGeometryWriter.java
@@ -1,25 +1,25 @@
-package mil.nga.giat.wkb.io;
+package mil.nga.wkb.io;
 
 import java.io.IOException;
 import java.nio.ByteOrder;
 
-import mil.nga.giat.wkb.geom.CircularString;
-import mil.nga.giat.wkb.geom.CompoundCurve;
-import mil.nga.giat.wkb.geom.Curve;
-import mil.nga.giat.wkb.geom.CurvePolygon;
-import mil.nga.giat.wkb.geom.Geometry;
-import mil.nga.giat.wkb.geom.GeometryCollection;
-import mil.nga.giat.wkb.geom.GeometryType;
-import mil.nga.giat.wkb.geom.LineString;
-import mil.nga.giat.wkb.geom.MultiLineString;
-import mil.nga.giat.wkb.geom.MultiPoint;
-import mil.nga.giat.wkb.geom.MultiPolygon;
-import mil.nga.giat.wkb.geom.Point;
-import mil.nga.giat.wkb.geom.Polygon;
-import mil.nga.giat.wkb.geom.PolyhedralSurface;
-import mil.nga.giat.wkb.geom.TIN;
-import mil.nga.giat.wkb.geom.Triangle;
-import mil.nga.giat.wkb.util.WkbException;
+import mil.nga.wkb.geom.CircularString;
+import mil.nga.wkb.geom.CompoundCurve;
+import mil.nga.wkb.geom.Curve;
+import mil.nga.wkb.geom.CurvePolygon;
+import mil.nga.wkb.geom.Geometry;
+import mil.nga.wkb.geom.GeometryCollection;
+import mil.nga.wkb.geom.GeometryType;
+import mil.nga.wkb.geom.LineString;
+import mil.nga.wkb.geom.MultiLineString;
+import mil.nga.wkb.geom.MultiPoint;
+import mil.nga.wkb.geom.MultiPolygon;
+import mil.nga.wkb.geom.Point;
+import mil.nga.wkb.geom.Polygon;
+import mil.nga.wkb.geom.PolyhedralSurface;
+import mil.nga.wkb.geom.TIN;
+import mil.nga.wkb.geom.Triangle;
+import mil.nga.wkb.util.WkbException;
 
 /**
  * Well Known Binary writer
